feat: add check for expect and rename file
