--- tests/lib/rules/no-multiple-expect-wait-for.test.ts
@@ -1,23 +1,0 @@
-import { createRuleTester } from '../test-utils';
-import rule, { RULE_NAME } from '../../../lib/rules/no-multiple-expect-wait-for';
-
-const ruleTester = createRuleTester({
-  ecmaFeatures: {
-    jsx: true,
-  },
-});
-
-ruleTester.run(RULE_NAME, rule, {
-  valid: [],
-  invalid: [
-    {
-      code: `
-        await waitFor(() => {
-          expect(a).toEqual('a')
-          expect(b).toEqual('b')
-        })
-      `,
-      errors: []
-    }
-  ]
-})
